replyModify 구현해야함

+++ src/com/sbs/java/blog/controller/ArticleController.java
@@ -2,7 +2,6 @@ package com.sbs.java.blog.controller;
 
 import java.sql.Connection;
 import java.util.List;
-import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -53,38 +52,45 @@ public class ArticleController extends Controller {
 
 		case "replyDelete":
 			return actionReplyDelete();
-			
+
 		case "replyModify":
 			return actionReplyModify();
-			
+
 		case "doReplyModify":
 			return actionDoReplyModify();
 		}
 		return "";
 	}
-	
 
 	private String actionDoReplyModify() {
-		
-		articleService.modifyReply(replyId);
-
-		return "html:<script> alert('댓글이 수정되었습니다.'); location.replace('history.back()')</script>";
+		int replyId = Util.getInt(req, "id");
+		String body = Util.getString(req, "body");
+		articleService.modifyReply(replyId, body);
+		return "html:<script> alert('댓글이 수정되었습니다.'); location.replace(history.back())</script>";
 	}
 
 	private String actionReplyModify() {
+		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
 		int replyId = Util.getInt(req, "id");
-		
-		return "articleReply/articleReplyModify.jsp";
+		ArticleReply reply = articleService.getArticleReplyByReplyId(replyId);
+		int replyMemeberId = Integer.parseInt(reply.getMemberId());
+		if (loginedMemberId == replyMemeberId) {
+			return "articleReply/replyModify.jsp";
+		}
+		return "html:<script> alert('권한이 없습니다.'); location.replace(history.back());</script>";
 	}
 
 	private String actionReplyDelete() {
 		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
-		
 		int replyId = Util.getInt(req, "id");
-		
-		articleService.deleteReply(replyId);
+		ArticleReply reply = articleService.getArticleReplyByReplyId(replyId);
+		int replyMemeberId = Integer.parseInt(reply.getMemberId());
 
-		return "html:<script> alert('댓글이 삭제되었습니다.'); location.replace('history.back()')</script>";
+		if (loginedMemberId == replyMemeberId) {
+			articleService.deleteReply(replyId);
+			return "html:<script> alert('댓글이 삭제되었습니다.'); location.replace('history.back()')</script>";
+		}
+		return "html:<script> alert('권한이 없습니다.'); location.replace('history.back()')</script>";
 	}
 
 	private String actionWriteArticleReply() {
@@ -130,9 +136,9 @@ public class ArticleController extends Controller {
 		if (Util.isNum(req, "id") == false) {
 			return "html:id를 숫자로 입력해 주세요.";
 		}
-		int loginedMemberId = (int) req.getAttribute("loginedMemberId"); 
+		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
 		int id = Util.getInt(req, "id");
-		Article article = articleService.detail(id,loginedMemberId);
+		Article article = articleService.detail(id, loginedMemberId);
 		System.out.println(article);
 		req.setAttribute("article", article);
 		return "article/modify.jsp";
@@ -165,9 +171,9 @@ public class ArticleController extends Controller {
 
 		int id = Util.getInt(req, "id");
 		articleService.increaseHit(id);
-		
+
 		int loginedMemberId = (int) req.getAttribute("loginedMemberId");
-		Article article = articleService.detail(id,loginedMemberId);
+		Article article = articleService.detail(id, loginedMemberId);
 		req.setAttribute("article", article);
 
 		int memberId = article.getMemberId();
@@ -181,9 +187,9 @@ public class ArticleController extends Controller {
 	}
 
 	private String actionList() {
-		
+
 //		long startTime = System.nanoTime();
-		
+
 		int page = 1;
 
 		if (!Util.empty(req, "page") && Util.isNum(req, "page")) {
@@ -228,7 +234,7 @@ public class ArticleController extends Controller {
 				searchKeyword);
 
 		req.setAttribute("articles", articles);
-		
+
 //		long endTime= System.nanoTime();
 //		long estimatedTime = endTime - startTime;
 //		//nano seconds to seconds
